[IPS-1805] refactor: Rename IPSApiWrapper to IPSRequests
